search bar functionality added, dark mode layout started
