--- src/components/NewFeedbackButton.js
@@ -1,12 +1,0 @@
-import "../styles/NewFeedbackButton.css";
-
-import React from "react";
-import Button from "react-bootstrap/Button";
-
-class NewFeedbackButton extends React.Component {
-  render() {
-    return <Button id="button-new">NEW FEEDBACK</Button>;
-  }
-}
-
-export default NewFeedbackButton;
