--- src/components/UserFeedbackCard.js
@@ -1,90 +1,0 @@
-import "../styles/UserFeedbackCard.css";
-
-import React from "react";
-import Container from "react-bootstrap/Container";
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
-import moment from "moment";
-
-class Report extends React.Component {
-  render() {
-    const formattedDate = moment(this.props.feedback.timestamp).format(
-      "MMMM Do YYYY, h:mm a"
-    );
-    const status = this.props.feedback.status;
-    let output;
-
-    if (status === "New") {
-      output = (
-        <button id="new-badge" disabled={true}>
-          New
-        </button>
-      );
-    }
-    if (status === "In Progress") {
-      output = (
-        <button id="in-progress-badge" disabled={true}>
-          In Progress
-        </button>
-      );
-    }
-    if (status === "Completed") {
-      output = (
-        <button id="completed-badge" disabled={true}>
-          Completed
-        </button>
-      );
-    }
-    if (status === "Closed") {
-      output = (
-        <button id="closed-badge" disabled={true}>
-          Closed
-        </button>
-      );
-    }
-    return (
-      <div className="user-feedback-card">
-        <Container>
-          <Row>
-            <Col xs={9} id="col1-card">
-              <p id="user-feedback-card-subject">
-                {this.props.feedback.subject}
-              </p>
-              <p id="user-feedback-card-tag">{this.props.feedback.project}</p>
-            </Col>
-            <Col xs={3} id="col2-card">
-              {output}
-            </Col>
-          </Row>
-        </Container>
-        <hr id="user-feedback-card-hr" />
-        <Container>
-          <Row>
-            <Col xs={8} id="col1-sub">
-              <p id="user-feedback-card-description">
-                {this.props.feedback.body}
-              </p>
-            </Col>
-            <Col xs={4} id="col2-sub">
-              <h5 id="user-votes">USER VOTES</h5>
-              <p id="upvote-count">{this.props.feedback.upvote}</p>
-            </Col>
-          </Row>
-        </Container>
-        <Container>
-          <Row>
-            <Col xs={8}>
-              <a id="attachment-link" href="attachment-url">
-                View attachments
-              </a>
-            </Col>
-            <Col xs={4}>
-              <p id="user-feedback-card-date">{formattedDate}</p>
-            </Col>
-          </Row>
-        </Container>
-      </div>
-    );
-  }
-}
-export default Report;
