--- src/styles/Filter.css
@@ -1,17 +1,0 @@
-#filter_text {
-  font-family: "Roboto Slab", serif;
-  font-size: 36px;
-  font-weight: bold;
-}
-
-.filter_headers_text {
-  font-family: "Roboto", sans-serif;
-  font-size: 24px;
-}
-
-#dropdown-basic-button {
-  margin-bottom: 20px;
-  border: 1px solid #cccccc;
-  box-sizing: border-box;
-  border-radius: 3px;
-}
