--- src/styles/Navigation.css
@@ -1,67 +1,0 @@
-.townhall_style {
-  font-family: Roboto Slab;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 24px;
-  line-height: 32px;
-  letter-spacing: 0.01em;
-}
-
-#town_style {
-  color: #2e5876;
-}
-
-#hall_style {
-  color: #51e2cd;
-}
-
-#navbar {
-  box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.1);
-  margin-bottom: 40px;
-  height: 75px;
-}
-
-#search {
-  background-color: #f3f3f3;
-  border-radius: 3px;
-  border-color: #f3f3f3;
-  font-family: Roboto;
-  font-size: 18px;
-  line-height: 21px;
-  color: #848484;
-  width: 281px;
-  height: 39px;
-}
-
-.dev-toggle.react-toggle--checked .react-toggle-track {
-  background-color: #2e5876;
-}
-
-.dev-toggle.react-toggle--checked:hover:not(.react-toggle--disabled)
-  .react-toggle-track {
-  background-color: #2e5876;
-}
-
-.react-toggle-thumb {
-  border: 1px solid rgb(177, 177, 177);
-}
-
-.dev-toggle.react-toggle--checked .react-toggle-thumb {
-  border-color: #2e5876;
-}
-
-.react-toggle-track {
-  background-color: rgb(177, 177, 177);
-}
-
-.react-toggle:hover:not(.react-toggle--disabled) .react-toggle-track {
-  background-color: rgb(177, 177, 177);
-}
-
-#toggle-label {
-  padding-right: 10px;
-  font-family: Roboto;
-  font-size: 18px;
-  line-height: 21px;
-  color: #848484;
-}
