--- src/styles/UserFeedbackCard.css
@@ -1,141 +1,0 @@
-.user-feedback-card {
-  width: 645px;
-  height: 234px;
-  background: #ffffff;
-  box-shadow: 0px 6px 10px rgba(0, 0, 0, 0.1);
-  border-radius: 3px;
-  margin: 0 auto;
-  margin-bottom: 20px;
-}
-
-#user-feedback-card-subject {
-  margin-top: 22.5px;
-  font-family: Roboto;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: #555555;
-}
-
-#user-feedback-card-tag {
-  margin-top: -12px;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 16px;
-  color: #848484;
-}
-
-#new-badge {
-  margin-top: 22.5px;
-  width: 128px;
-  height: 44px;
-  background-color: darkslategrey;
-  border-radius: 22px;
-  border: none;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: white;
-}
-
-#in-progress-badge {
-  margin-top: 22.5px;
-  width: 128px;
-  height: 44px;
-  background-color: darkslategrey;
-  border-radius: 22px;
-  border: none;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: white;
-}
-
-#completed-badge {
-  margin-top: 22.5px;
-  width: 128px;
-  height: 44px;
-  background-color: darkslategrey;
-  border-radius: 22px;
-  border: none;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: white;
-}
-
-#closed-badge {
-  margin-top: 22.5px;
-  width: 128px;
-  height: 44px;
-  background-color: darkslategrey;
-  border-radius: 22px;
-  border: none;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: white;
-}
-
-#user-feedback-card-hr {
-  width: 645px;
-  height: 0px;
-  border: 0.5px solid #e5e5e5;
-  margin-top: 5px;
-}
-
-#user-feedback-card-description {
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 21px;
-  color: #555555;
-}
-
-#user-votes {
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 16px;
-  color: #848484;
-}
-
-#upvote-count {
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 36px;
-  line-height: 42px;
-  color: #555555;
-}
-
-#attachment-link {
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 16px;
-  color: #848484;
-  text-decoration: underline;
-}
-
-#user-feedback-card-date {
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 14px;
-  line-height: 16px;
-  color: #848484;
-}
